Fix mentions coming from Matrix
