+++ server/matrix_util_test.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/golang/mock/gomock"
 	"github.com/mattermost/mattermost-plugin-matrix-bridge/server/mocks"
-	"github.com/mattermost/mattermost/server/public/model"
 	"github.com/stretchr/testify/require"
 )
 
@@ -425,520 +424,6 @@ func TestConvertHTMLToMarkdownError(t *testing.T) {
 	// Library handles malformed HTML without errors, so no warning expected
 	if len(mockLogger.logs) > 0 {
 		t.Errorf("convertHTMLToMarkdown should not log warnings for this input, but got: %v", mockLogger.logs)
-	}
-}
-
-func TestExtractMentionedUsers(t *testing.T) {
-	tests := []struct {
-		name     string
-		event    MatrixEvent
-		expected []string
-	}{
-		{
-			name: "no mentions field",
-			event: MatrixEvent{
-				EventID: "event1",
-				Content: map[string]any{},
-			},
-			expected: nil,
-		},
-		{
-			name: "empty mentions",
-			event: MatrixEvent{
-				EventID: "event2",
-				Content: map[string]any{
-					"m.mentions": map[string]any{},
-				},
-			},
-			expected: nil,
-		},
-		{
-			name: "single mention",
-			event: MatrixEvent{
-				EventID: "event3",
-				Content: map[string]any{
-					"m.mentions": map[string]any{
-						"user_ids": []any{"@alice:example.com"},
-					},
-				},
-			},
-			expected: []string{"@alice:example.com"},
-		},
-		{
-			name: "multiple mentions",
-			event: MatrixEvent{
-				EventID: "event4",
-				Content: map[string]any{
-					"m.mentions": map[string]any{
-						"user_ids": []any{"@alice:example.com", "@bob:example.com"},
-					},
-				},
-			},
-			expected: []string{"@alice:example.com", "@bob:example.com"},
-		},
-		{
-			name: "invalid mentions format",
-			event: MatrixEvent{
-				EventID: "event5",
-				Content: map[string]any{
-					"m.mentions": "invalid",
-				},
-			},
-			expected: nil,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Create a minimal plugin instance for testing
-			ctrl := gomock.NewController(t)
-			defer ctrl.Finish()
-			mockAPI := mocks.NewMockAPI(ctrl)
-
-			plugin := setupPluginForTestWithLogger(t, mockAPI)
-
-			result := plugin.extractMentionedUsers(tt.event)
-
-			if len(result) != len(tt.expected) {
-				t.Errorf("extractMentionedUsers() returned %d users, want %d", len(result), len(tt.expected))
-				return
-			}
-
-			for i, expected := range tt.expected {
-				if result[i] != expected {
-					t.Errorf("extractMentionedUsers()[%d] = %q, want %q", i, result[i], expected)
-				}
-			}
-		})
-	}
-}
-
-func TestReplaceMatrixMentionHTML(t *testing.T) {
-	tests := []struct {
-		name               string
-		htmlContent        string
-		matrixUserID       string
-		mattermostUsername string
-		expected           string
-		setup              func(*mocks.MockAPI)
-	}{
-		{
-			name:               "simple mention",
-			htmlContent:        `Hello <a href="https://matrix.to/#/@alice:example.com">Alice</a>!`,
-			matrixUserID:       "@alice:example.com",
-			mattermostUsername: "alice",
-			expected:           "Hello @alice!",
-		},
-		{
-			name:               "mention with display name",
-			htmlContent:        `Hey <a href="https://matrix.to/#/@bob:server.org">Bob Smith</a>, how are you?`,
-			matrixUserID:       "@bob:server.org",
-			mattermostUsername: "bobsmith",
-			expected:           "Hey @bobsmith, how are you?",
-		},
-		{
-			name:               "multiple mentions of same user",
-			htmlContent:        `<a href="https://matrix.to/#/@alice:example.com">Alice</a> and <a href="https://matrix.to/#/@alice:example.com">Alice again</a>`,
-			matrixUserID:       "@alice:example.com",
-			mattermostUsername: "alice",
-			expected:           "@alice and @alice",
-		},
-		{
-			name:               "no mention links",
-			htmlContent:        "Just plain text",
-			matrixUserID:       "@alice:example.com",
-			mattermostUsername: "alice",
-			expected:           "Just plain text",
-		},
-		{
-			name:               "mention with different formatting",
-			htmlContent:        `<a href='https://matrix.to/#/@user:example.com' class="mention">User Name</a>`,
-			matrixUserID:       "@user:example.com",
-			mattermostUsername: "username",
-			expected:           "@username",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Create a minimal plugin instance for testing
-			ctrl := gomock.NewController(t)
-			defer ctrl.Finish()
-			mockAPI := mocks.NewMockAPI(ctrl)
-
-			plugin := setupPluginForTestWithLogger(t, mockAPI)
-
-			result := plugin.replaceMatrixMentionHTML(tt.htmlContent, tt.matrixUserID, tt.mattermostUsername)
-
-			if result != tt.expected {
-				t.Errorf("replaceMatrixMentionHTML() = %q, want %q", result, tt.expected)
-			}
-		})
-	}
-}
-
-func TestGetMattermostUsernameFromMatrix(t *testing.T) {
-	tests := []struct {
-		name         string
-		matrixUserID string
-		expected     string
-		setup        func(*mocks.MockAPI, *mocks.MockKVStore)
-	}{
-		{
-			name:         "user not found in kvstore",
-			matrixUserID: "@alice:example.com",
-			expected:     "",
-			setup: func(_ *mocks.MockAPI, mockKV *mocks.MockKVStore) {
-				mockKV.EXPECT().Get("matrix_user_@alice:example.com").Return(nil, &model.AppError{Message: "Not found"})
-			},
-		},
-		{
-			name:         "user found but Mattermost user doesn't exist",
-			matrixUserID: "@bob:example.com",
-			expected:     "",
-			setup: func(mockAPI *mocks.MockAPI, mockKV *mocks.MockKVStore) {
-				mockKV.EXPECT().Get("matrix_user_@bob:example.com").Return([]byte("user123"), nil)
-				mockAPI.EXPECT().GetUser("user123").Return(nil, &model.AppError{Message: "User not found"})
-			},
-		},
-		{
-			name:         "user found successfully",
-			matrixUserID: "@charlie:example.com",
-			expected:     "charlie_mm",
-			setup: func(mockAPI *mocks.MockAPI, mockKV *mocks.MockKVStore) {
-				mockKV.EXPECT().Get("matrix_user_@charlie:example.com").Return([]byte("user456"), nil)
-				user := &model.User{
-					Id:       "user456",
-					Username: "charlie_mm",
-				}
-				mockAPI.EXPECT().GetUser("user456").Return(user, nil)
-			},
-		},
-		{
-			name:         "ghost user found successfully",
-			matrixUserID: "@_mattermost_yeqo3irkujdstfmbnkx46bbhuw:synapse-mydomain.com",
-			expected:     "doug_lauder",
-			setup: func(mockAPI *mocks.MockAPI, _ *mocks.MockKVStore) {
-				user := &model.User{
-					Id:       "yeqo3irkujdstfmbnkx46bbhuw",
-					Username: "doug_lauder",
-				}
-				mockAPI.EXPECT().GetUser("yeqo3irkujdstfmbnkx46bbhuw").Return(user, nil)
-			},
-		},
-		{
-			name:         "ghost user but Mattermost user doesn't exist",
-			matrixUserID: "@_mattermost_nonexistentuser:example.com",
-			expected:     "",
-			setup: func(mockAPI *mocks.MockAPI, _ *mocks.MockKVStore) {
-				mockAPI.EXPECT().GetUser("nonexistentuser").Return(nil, &model.AppError{Message: "User not found"})
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Create plugin instance with mocks
-			ctrl := gomock.NewController(t)
-			defer ctrl.Finish()
-			mockAPI := mocks.NewMockAPI(ctrl)
-			mockKV := mocks.NewMockKVStore(ctrl)
-
-			// Set up test-specific expectations
-			tt.setup(mockAPI, mockKV)
-
-			plugin := setupPluginForTestWithKVStore(t, mockAPI, mockKV)
-
-			result := plugin.getMattermostUsernameFromMatrix(tt.matrixUserID)
-
-			if result != tt.expected {
-				t.Errorf("getMattermostUsernameFromMatrix() = %q, want %q", result, tt.expected)
-			}
-		})
-	}
-}
-
-func TestProcessMatrixMentions(t *testing.T) {
-	tests := []struct {
-		name        string
-		htmlContent string
-		event       MatrixEvent
-		expected    string
-		setup       func(*mocks.MockAPI, *mocks.MockKVStore)
-	}{
-		{
-			name:        "no mentions in event",
-			htmlContent: `Hello <a href="https://matrix.to/#/@alice:example.com">Alice</a>!`,
-			event: MatrixEvent{
-				EventID: "event1",
-				Content: map[string]any{},
-			},
-			expected: `Hello <a href="https://matrix.to/#/@alice:example.com">Alice</a>!`,
-			setup: func(_ *mocks.MockAPI, _ *mocks.MockKVStore) {
-				// No expectations needed
-			},
-		},
-		{
-			name:        "single mention found and replaced",
-			htmlContent: `Hello <a href="https://matrix.to/#/@alice:example.com">Alice</a>!`,
-			event: MatrixEvent{
-				EventID: "event2",
-				Content: map[string]any{
-					"m.mentions": map[string]any{
-						"user_ids": []any{"@alice:example.com"},
-					},
-				},
-			},
-			expected: "Hello @alice_mm!",
-			setup: func(mockAPI *mocks.MockAPI, mockKV *mocks.MockKVStore) {
-				mockKV.EXPECT().Get("matrix_user_@alice:example.com").Return([]byte("user123"), nil)
-				user := &model.User{
-					Id:       "user123",
-					Username: "alice_mm",
-				}
-				mockAPI.EXPECT().GetUser("user123").Return(user, nil)
-			},
-		},
-		{
-			name:        "multiple mentions processed",
-			htmlContent: `<a href="https://matrix.to/#/@alice:example.com">Alice</a> and <a href="https://matrix.to/#/@bob:example.com">Bob</a>`,
-			event: MatrixEvent{
-				EventID: "event3",
-				Content: map[string]any{
-					"m.mentions": map[string]any{
-						"user_ids": []any{"@alice:example.com", "@bob:example.com"},
-					},
-				},
-			},
-			expected: "@alice_mm and @bob_mm",
-			setup: func(mockAPI *mocks.MockAPI, mockKV *mocks.MockKVStore) {
-				// Setup for Alice
-				mockKV.EXPECT().Get("matrix_user_@alice:example.com").Return([]byte("user123"), nil)
-				aliceUser := &model.User{
-					Id:       "user123",
-					Username: "alice_mm",
-				}
-				mockAPI.EXPECT().GetUser("user123").Return(aliceUser, nil)
-
-				// Setup for Bob
-				mockKV.EXPECT().Get("matrix_user_@bob:example.com").Return([]byte("user456"), nil)
-				bobUser := &model.User{
-					Id:       "user456",
-					Username: "bob_mm",
-				}
-				mockAPI.EXPECT().GetUser("user456").Return(bobUser, nil)
-			},
-		},
-		{
-			name:        "mention with no Mattermost mapping",
-			htmlContent: `Hello <a href="https://matrix.to/#/@unknown:example.com">Unknown</a>!`,
-			event: MatrixEvent{
-				EventID: "event4",
-				Content: map[string]any{
-					"m.mentions": map[string]any{
-						"user_ids": []any{"@unknown:example.com"},
-					},
-				},
-			},
-			expected: `Hello <a href="https://matrix.to/#/@unknown:example.com">Unknown</a>!`,
-			setup: func(_ *mocks.MockAPI, mockKV *mocks.MockKVStore) {
-				mockKV.EXPECT().Get("matrix_user_@unknown:example.com").Return(nil, &model.AppError{Message: "Not found"})
-			},
-		},
-		{
-			name:        "mixed content with mentions and regular links",
-			htmlContent: `Check out <a href="https://example.com">this link</a> and <a href="https://matrix.to/#/@alice:example.com">Alice</a>`,
-			event: MatrixEvent{
-				EventID: "event5",
-				Content: map[string]any{
-					"m.mentions": map[string]any{
-						"user_ids": []any{"@alice:example.com"},
-					},
-				},
-			},
-			expected: `Check out <a href="https://example.com">this link</a> and @alice_mm`,
-			setup: func(mockAPI *mocks.MockAPI, mockKV *mocks.MockKVStore) {
-				mockKV.EXPECT().Get("matrix_user_@alice:example.com").Return([]byte("user123"), nil)
-				user := &model.User{
-					Id:       "user123",
-					Username: "alice_mm",
-				}
-				mockAPI.EXPECT().GetUser("user123").Return(user, nil)
-			},
-		},
-		{
-			name:        "ghost user mention from real example",
-			htmlContent: `This is a test mention for <a href="https://matrix.to/#/@_mattermost_yeqo3irkujdstfmbnkx46bbhuw:synapse-mydomain.com">Doug Lauder</a> hope it works`,
-			event: MatrixEvent{
-				EventID: "event6",
-				Content: map[string]any{
-					"m.mentions": map[string]any{
-						"user_ids": []any{"@_mattermost_yeqo3irkujdstfmbnkx46bbhuw:synapse-mydomain.com"},
-					},
-				},
-			},
-			expected: `This is a test mention for @doug_lauder hope it works`,
-			setup: func(mockAPI *mocks.MockAPI, _ *mocks.MockKVStore) {
-				user := &model.User{
-					Id:       "yeqo3irkujdstfmbnkx46bbhuw",
-					Username: "doug_lauder",
-				}
-				mockAPI.EXPECT().GetUser("yeqo3irkujdstfmbnkx46bbhuw").Return(user, nil)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctrl := gomock.NewController(t)
-			defer ctrl.Finish()
-			mockAPI := mocks.NewMockAPI(ctrl)
-			mockKV := mocks.NewMockKVStore(ctrl)
-
-			tt.setup(mockAPI, mockKV)
-
-			plugin := setupPluginForTestWithKVStore(t, mockAPI, mockKV)
-
-			result := plugin.processMatrixMentions(tt.htmlContent, tt.event)
-
-			if result != tt.expected {
-				t.Errorf("processMatrixMentions() = %q, want %q", result, tt.expected)
-			}
-		})
-	}
-}
-
-func TestConvertHTMLToMarkdownWithMentions(t *testing.T) {
-	tests := []struct {
-		name        string
-		htmlContent string
-		event       MatrixEvent
-		expected    string
-		setup       func(*mocks.MockAPI, *mocks.MockKVStore)
-	}{
-		{
-			name:        "plain HTML without mentions",
-			htmlContent: "<strong>Hello</strong> <em>world</em>",
-			event: MatrixEvent{
-				EventID: "event1",
-				Content: map[string]any{},
-			},
-			expected: "**Hello** *world*",
-			setup: func(_ *mocks.MockAPI, _ *mocks.MockKVStore) {
-				// No expectations needed
-			},
-		},
-		{
-			name:        "HTML with mentions converted to markdown",
-			htmlContent: `<p>Hello <a href="https://matrix.to/#/@alice:example.com">Alice</a>, how are you?</p>`,
-			event: MatrixEvent{
-				EventID: "event2",
-				Content: map[string]any{
-					"m.mentions": map[string]any{
-						"user_ids": []any{"@alice:example.com"},
-					},
-				},
-			},
-			expected: "Hello @alice\\_mm, how are you?",
-			setup: func(mockAPI *mocks.MockAPI, mockKV *mocks.MockKVStore) {
-				mockKV.EXPECT().Get("matrix_user_@alice:example.com").Return([]byte("user123"), nil)
-				user := &model.User{
-					Id:       "user123",
-					Username: "alice_mm",
-				}
-				mockAPI.EXPECT().GetUser("user123").Return(user, nil)
-			},
-		},
-		{
-			name:        "complex HTML with mentions and formatting",
-			htmlContent: `<p><strong>Important:</strong> <a href="https://matrix.to/#/@alice:example.com">Alice</a> needs to see this <em>urgent</em> message!</p>`,
-			event: MatrixEvent{
-				EventID: "event3",
-				Content: map[string]any{
-					"m.mentions": map[string]any{
-						"user_ids": []any{"@alice:example.com"},
-					},
-				},
-			},
-			expected: "**Important:** @alice\\_mm needs to see this *urgent* message!",
-			setup: func(mockAPI *mocks.MockAPI, mockKV *mocks.MockKVStore) {
-				mockKV.EXPECT().Get("matrix_user_@alice:example.com").Return([]byte("user123"), nil)
-				user := &model.User{
-					Id:       "user123",
-					Username: "alice_mm",
-				}
-				mockAPI.EXPECT().GetUser("user123").Return(user, nil)
-			},
-		},
-		{
-			name:        "mentions with no Mattermost mapping remain as HTML",
-			htmlContent: `<p>Hello <a href="https://matrix.to/#/@unknown:example.com">Unknown User</a></p>`,
-			event: MatrixEvent{
-				EventID: "event4",
-				Content: map[string]any{
-					"m.mentions": map[string]any{
-						"user_ids": []any{"@unknown:example.com"},
-					},
-				},
-			},
-			expected: "Hello [Unknown User](https://matrix.to/#/@unknown:example.com)",
-			setup: func(_ *mocks.MockAPI, mockKV *mocks.MockKVStore) {
-				mockKV.EXPECT().Get("matrix_user_@unknown:example.com").Return(nil, &model.AppError{Message: "Not found"})
-			},
-		},
-		{
-			name:        "multiple mentions with mixed formatting",
-			htmlContent: `<ul><li><a href="https://matrix.to/#/@alice:example.com">Alice</a> - <strong>Team Lead</strong></li><li><a href="https://matrix.to/#/@bob:example.com">Bob</a> - <em>Developer</em></li></ul>`,
-			event: MatrixEvent{
-				EventID: "event5",
-				Content: map[string]any{
-					"m.mentions": map[string]any{
-						"user_ids": []any{"@alice:example.com", "@bob:example.com"},
-					},
-				},
-			},
-			expected: "- @alice\\_mm - **Team Lead**\n- @bob\\_mm - *Developer*",
-			setup: func(mockAPI *mocks.MockAPI, mockKV *mocks.MockKVStore) {
-				// Setup for Alice
-				mockKV.EXPECT().Get("matrix_user_@alice:example.com").Return([]byte("user123"), nil)
-				aliceUser := &model.User{
-					Id:       "user123",
-					Username: "alice_mm",
-				}
-				mockAPI.EXPECT().GetUser("user123").Return(aliceUser, nil)
-
-				// Setup for Bob
-				mockKV.EXPECT().Get("matrix_user_@bob:example.com").Return([]byte("user456"), nil)
-				bobUser := &model.User{
-					Id:       "user456",
-					Username: "bob_mm",
-				}
-				mockAPI.EXPECT().GetUser("user456").Return(bobUser, nil)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctrl := gomock.NewController(t)
-			defer ctrl.Finish()
-			mockAPI := mocks.NewMockAPI(ctrl)
-			mockKV := mocks.NewMockKVStore(ctrl)
-
-			tt.setup(mockAPI, mockKV)
-
-			plugin := setupPluginForTestWithKVStore(t, mockAPI, mockKV)
-
-			result := plugin.convertHTMLToMarkdownWithMentions(tt.htmlContent, tt.event)
-
-			// Normalize whitespace for comparison
-			result = strings.TrimSpace(result)
-			expected := strings.TrimSpace(tt.expected)
-
-			if result != expected {
-				t.Errorf("convertHTMLToMarkdownWithMentions() = %q, want %q", result, expected)
-			}
-		})
 	}
 }
